refactor: 이름 변경 및 패키지 이동
